requirements.txt
flask
requests
pyjwt
flask-SQLAlchemy
flask_validator
psycopg2-binary
python-dotenv
pkg-resources
PyYAML